// src/ts/types/GithubUser.ts
export interface GithubUser {
  avatarUrl: string;
  bio: string;
  blog: string;
  company: string;
  createdAt: string;
  followers: number;
  followersUrl: string;
  following: number;
  followingUrl: string;
  id: number;
  location: string;
  name: string;
  publicRepos: number;
  reposUrl: string;
  twitterUsername: string;
  url: string;
  username: string;
}